Provide types to Property Options via `PluginOptionType`
