+++ core/src/modules/optionHelper.ts
@@ -3,13 +3,14 @@ import {
   GrouparooPlugin,
   PluginConnection,
   PluginApp,
+  AppOption,
 } from "../classes/plugin";
 import { Option } from "./../models/Option";
 import { Source } from "./../models/Source";
 import { Destination } from "./../models/Destination";
 import { Schedule, SimpleScheduleOptions } from "./../models/Schedule";
 import { Property, SimplePropertyOptions } from "../models/Property";
-import { App, AppOption } from "./../models/App";
+import { App } from "./../models/App";
 import { LoggedModel } from "../classes/loggedModel";
 import { LockableHelper } from "./lockableHelper";
 import { plural } from "pluralize";
